perf: reduce memory use in storing eols

+++ tools/Console/Program.cs
@@ -38,7 +38,7 @@ namespace FlameCsv.Console
                 var rb = new RecordBuffer();
                 rb.GetFieldArrayRef() = new uint[24 * 65535];
                 rb.GetQuoteArrayRef() = new byte[24 * 65535];
-                rb.GetEolArrayRef() = new int[65535 + 256];
+                rb.GetEolArrayRef() = new ushort[65535 + 256];
 
                 // new SimdTokenizer<byte, NewlineCRLF>(CsvOptions<byte>.Default)
                 //     .Tokenize(rb, Encoding.UTF8.GetBytes(Encoding.UTF8.GetString(byteArray).ReplaceLineEndings("\r\n")));
